# dispatcher/backend/docs/openapi_v1.yaml
openapi: 3.0.0
servers:
  - description: production server
    url: https://api.farm.openzim.org/v1
info:
  description: |
    Zimfarm is openZIM's automated & distributed ZIM files creation engine.
    openZIM periodically publishes updated ZIM files for a large variety of content.

    This API provides public access to the scheduled ZIM files to create as well as tracking of its tasks.

    This API is also used by _workers_ to fetch, process and update those tasks.
  version: 1.0.0
  title: Zimfarm API
  contact:
    email: contact+zimfarm@kiwix.org
    url: https://kiwix.org
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
...
paths:
...
  /users/:
    get:
      tags:
      - restricted
      summary: list users
      operationId: listUsers
      description: Query the Users database
      security:
        - token: []
        - oauth: []
      parameters:
      - $ref: '#/components/parameters/SkipParameter'
      - $ref: '#/components/parameters/LimitParameter'
      responses:
        200:
          description: List of Users matching criteria
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/MetaProperty'
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserItem'
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
        401:
          description: Unauthorized
    post:
      tags:
      - restricted
      summary: create user
      operationId: addUser
      description: |
        Create a new User with specific scope and credentials.
        SSH Authentication requires an additional key upload step.
        API user needs the `users.create` scope for this endpoint.
      security:
        - token: []
        - oauth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPayload'
            examples:
              username:
                summary: John Doe
                value: {"username": "john", "password": "my-long-password", "email": "john@kiwix.org", "scope": {"schedules": {"read": true}}}
      responses:
        201:
          description: User Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    $ref: '#/components/schemas/MongoId'
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
        401:
          description: Unauthorized
  /users/{username}:
    get:
      tags:
      - restricted
      summary: get user details
      operationId: getUser
      description: Retrieve a User's details
      security:
        - token: []
        - ouath: []
      parameters:
      - $ref: '#/components/parameters/UsernameParameter'
      responses:
        200:
          description: User Record
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    $ref: '#/components/schemas/Username'
                  email:
                    $ref: '#/components/schemas/UserEmail'
                  scope:
                    $ref: '#/components/schemas/Scope'
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
    patch:
      tags:
      - restricted
      summary: update user details
      operationId: updateUser
      description: Change a User's details
      security:
        - token: []
        - ouath: []
      parameters:
      - $ref: '#/components/parameters/UsernameParameter'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                email:
                  $ref: '#/components/schemas/Email'
                role:
                  $ref: '#/components/schemas/Role'
      responses:
        204:
          description: User Updated
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
    delete:
      tags:
      - restricted
      summary: delete user
      operationId: deleteUser
      description: Remove a User
      parameters:
      - $ref: '#/components/parameters/UsernameParameter'
      security:
        - token: []
        - oauth: []
      responses:
        200:
          description: User Deleted
        401:
          description: Unauthorized
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    token:
      type: apiKey
      description: API Token
      name: token
      in: header
    oauth:
      type: oauth2
      description: OAuth2
      flows:
        password:
          tokenUrl: https://api.farm.openzim.org/v1/auth/oauth2
          refreshUrl: https://api.farm.openzim.org/v1/auth/oauth2
  parameters:
    SkipParameter:
      in: query
      name: skip
      description: number of records to skip for pagination (yay)
      required: false
      schema:
        $ref: '#/components/schemas/SkipProperty'
    LimitParameter:
      in: query
      name: limit
      description: maximum number of records to return
      required: false
      schema:
        $ref: '#/components/schemas/LimitProperty'
...
    UsernameParameter:
      in: path
      required: true
      name: username
      schema:
        $ref: '#/components/schemas/Username'
...
  schemas:
...
    UserItem:
      type: object
      required:
      - username
      - email
      properties:
        username:
          $ref: '#/components/schemas/Username'
        email:
          $ref: '#/components/schemas/UserEmail'
    UserPayload:
      type: object
      additionalProperties: false
      required:
      - username
      - password
      - email
      - scope
      properties:
        username:
          $ref: '#/components/schemas/Username'
        password:
          type: string
          format: password
          minLength: 6
        email:
          $ref: '#/components/schemas/UserEmail'
        scope:
          $ref: '#/components/schemas/Scope'
    Scope:
      type: object
      properties:
        schedules:
          $ref: '#/components/schemas/ScopeItem'
        tasks:
          $ref: '#/components/schemas/ScopeItem'
        users:
          type: object
          properties:
            read:
              $ref: '#/components/schemas/BooleanProperty'
            create:
              $ref: '#/components/schemas/BooleanProperty'
            update:
              $ref: '#/components/schemas/BooleanProperty'
            delete:
              $ref: '#/components/schemas/BooleanProperty'
            reset_password:
              $ref: '#/components/schemas/BooleanProperty'
            keys:
              type: object
              properties:
                read:
                  $ref: '#/components/schemas/BooleanProperty'
                delete:
                  $ref: '#/components/schemas/BooleanProperty'
    ScopeItem:
      type: object
      properties:
        read:
          $ref: '#/components/schemas/BooleanProperty'
        create:
          $ref: '#/components/schemas/BooleanProperty'
        update:
          $ref: '#/components/schemas/BooleanProperty'
        delete:
          $ref: '#/components/schemas/BooleanProperty'
    BooleanProperty:
      type: boolean
      example: true
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Not Found
    InputError:
      type: object
      properties:
        message:
          oneOf:
            - type: object
              example: {"schedule_name": "required field"}
            - type: string
              example: Invalid signature format (not base64)
...
    MetaProperty:
      type: object
      properties:
        count:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        skip:
          type: integer
          example: 0
    LimitProperty:
      type: integer
      format: int32
      minimum: 0
      example: 10
      default: 10
    SkipProperty:
      type: integer
      format: int32
      minimum: 0
      example: 0
      default: 0
    MongoId:
      type: string
      format: uuid
      example: 5dbc599b3b61e21c362703e4
    Username:
      type: string
      example: john
      minLength: 1
    UserEmail:
      type: string
      format: email
      example: john@doe.org
...
    Email:
      type: string
      format: email
      example: user@kiwix.org
    Role:
      type: string
      example: admin
      enum:
        - editor
        - manager
        - admin
        - worker
        - processor